requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
argon2-cffi==21.3.0
    # via notebook
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
asttokens==2.0.5
    # via stack-data
attrs==21.4.0
    # via jsonschema
backcall==0.2.0
    # via ipython
beautifulsoup4==4.11.1
    # via nbconvert
bleach==5.0.0
    # via nbconvert
certifi==2021.10.8
    # via
    #   elastic-transport
    #   requests
cffi==1.15.0
    # via argon2-cffi-bindings
charset-normalizer==2.0.12
    # via requests
click==8.1.2
    # via
    #   nltk
    #   sacremoses
commonmark==0.9.1
    # via rich
cycler==0.11.0
    # via matplotlib
debugpy==1.6.0
    # via ipykernel
decorator==5.1.1
    # via ipython
defusedxml==0.7.1
    # via nbconvert
eland[pytorch]==8.1.0
    # via -r requirements.in
elastic-transport==8.1.2
    # via elasticsearch
elasticsearch==8.1.3
    # via eland
entrypoints==0.4
    # via
    #   jupyter-client
    #   nbconvert
executing==0.8.3
    # via stack-data
fastjsonschema==2.15.3
    # via nbformat
filelock==3.6.0
    # via
    #   huggingface-hub
    #   transformers
fonttools==4.33.3
    # via matplotlib
huggingface-hub==0.5.1
    # via
    #   sentence-transformers
    #   transformers
idna==3.3
    # via requests
ipykernel==6.13.0
    # via
    #   ipywidgets
    #   jupyter
    #   jupyter-console
    #   notebook
    #   qtconsole
ipython==8.2.0
    # via
    #   ipykernel
    #   ipywidgets
    #   jupyter-console
ipython-genutils==0.2.0
    # via
    #   ipywidgets
    #   notebook
    #   qtconsole
ipywidgets==7.7.0
    # via jupyter
jedi==0.18.1
    # via ipython
jinja2==3.1.1
    # via
    #   nbconvert
    #   notebook
joblib==1.1.0
    # via
    #   nltk
    #   sacremoses
    #   scikit-learn
jsonschema==4.4.0
    # via nbformat
jupyter==1.0.0
    # via -r requirements.in
jupyter-client==7.3.0
    # via
    #   ipykernel
    #   jupyter-console
    #   nbclient
    #   notebook
    #   qtconsole
jupyter-console==6.4.3
    # via jupyter
jupyter-core==4.10.0
    # via
    #   jupyter-client
    #   nbconvert
    #   nbformat
    #   notebook
    #   qtconsole
jupyterlab-pygments==0.2.2
    # via nbconvert
jupyterlab-widgets==1.1.0
    # via ipywidgets
kiwisolver==1.4.2
    # via matplotlib
markupsafe==2.1.1
    # via
    #   jinja2
    #   nbconvert
matplotlib==3.5.1
    # via eland
matplotlib-inline==0.1.3
    # via
    #   ipykernel
    #   ipython
mistune==0.8.4
    # via nbconvert
nbclient==0.6.0
    # via nbconvert
nbconvert==6.5.0
    # via
    #   jupyter
    #   notebook
nbformat==5.3.0
    # via
    #   ipywidgets
    #   nbclient
    #   nbconvert
    #   notebook
nest-asyncio==1.5.5
    # via
    #   ipykernel
    #   jupyter-client
    #   nbclient
    #   notebook
nltk==3.7
    # via sentence-transformers
notebook==6.4.11
    # via
    #   jupyter
    #   widgetsnbextension
numpy==1.22.3
    # via
    #   eland
    #   matplotlib
    #   pandas
    #   scikit-learn
    #   scipy
    #   sentence-transformers
    #   torchvision
    #   transformers
packaging==21.3
    # via
    #   huggingface-hub
    #   ipykernel
    #   matplotlib
    #   nbconvert
    #   qtpy
    #   transformers
pandas==1.3.5
    # via eland
pandocfilters==1.5.0
    # via nbconvert
parso==0.8.3
    # via jedi
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pillow==9.1.0
    # via
    #   matplotlib
    #   torchvision
prometheus-client==0.14.1
    # via notebook
prompt-toolkit==3.0.29
    # via
    #   ipython
    #   jupyter-console
psutil==5.9.0
    # via ipykernel
ptyprocess==0.7.0
    # via
    #   pexpect
    #   terminado
pure-eval==0.2.2
    # via stack-data
pycparser==2.21
    # via cffi
pygments==2.12.0
    # via
    #   ipython
    #   jupyter-console
    #   nbconvert
    #   qtconsole
    #   rich
pyparsing==3.0.8
    # via
    #   matplotlib
    #   packaging
pyrsistent==0.18.1
    # via jsonschema
python-dateutil==2.8.2
    # via
    #   jupyter-client
    #   matplotlib
    #   pandas
pytz==2022.1
    # via pandas
pyyaml==6.0
    # via
    #   huggingface-hub
    #   transformers
pyzmq==22.3.0
    # via
    #   jupyter-client
    #   notebook
    #   qtconsole
qtconsole==5.3.0
    # via jupyter
qtpy==2.0.1
    # via qtconsole
regex==2022.4.24
    # via
    #   nltk
    #   sacremoses
    #   transformers
requests==2.27.1
    # via
    #   huggingface-hub
    #   torchvision
    #   transformers
rich==12.3.0
    # via -r requirements.in
sacremoses==0.0.49
    # via transformers
scikit-learn==1.0.2
    # via sentence-transformers
scipy==1.8.0
    # via
    #   scikit-learn
    #   sentence-transformers
send2trash==1.8.0
    # via notebook
sentence-transformers==2.2.0
    # via eland
sentencepiece==0.1.96
    # via sentence-transformers
six==1.16.0
    # via
    #   asttokens
    #   bleach
    #   python-dateutil
    #   sacremoses
soupsieve==2.3.2.post1
    # via beautifulsoup4
stack-data==0.2.0
    # via ipython
terminado==0.13.3
    # via notebook
threadpoolctl==3.1.0
    # via scikit-learn
tinycss2==1.1.1
    # via nbconvert
tokenizers==0.12.1
    # via transformers
torch==1.11.0
    # via
    #   eland
    #   sentence-transformers
    #   torchvision
    #   transformers
torchvision==0.12.0
    # via sentence-transformers
tornado==6.1
    # via
    #   ipykernel
    #   jupyter-client
    #   notebook
    #   terminado
tqdm==4.64.0
    # via
    #   huggingface-hub
    #   nltk
    #   sacremoses
    #   sentence-transformers
    #   transformers
traitlets==5.1.1
    # via
    #   ipykernel
    #   ipython
    #   ipywidgets
    #   jupyter-client
    #   jupyter-core
    #   matplotlib-inline
    #   nbclient
    #   nbconvert
    #   nbformat
    #   notebook
    #   qtconsole
transformers[torch]==4.18.0
    # via
    #   eland
    #   sentence-transformers
typing-extensions==4.2.0
    # via
    #   huggingface-hub
    #   torch
    #   torchvision
urllib3==1.26.9
    # via
    #   elastic-transport
    #   requests
wcwidth==0.2.5
    # via prompt-toolkit
webencodings==0.5.1
    # via
    #   bleach
    #   tinycss2
widgetsnbextension==3.6.0
    # via ipywidgets

# The following packages are considered to be unsafe in a requirements file:
# setuptools
